requirements.txt
pyperclip>=1.8.2
flask>=2.3.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0

